# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Keeppix API",
    "description": "API di Keeppix. Contratto congelato: solo aggiunte entro /api/v1.",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/assets/{id}": {
      "get": {
        "tags": [
          "timeline"
        ],
        "summary": "Get an asset",
        "operationId": "assets_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id asset",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vista pubblica dell'asset",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetView"
                }
              }
            }
          },
          "400": {
            "description": "Id non valido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Asset non visibile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Asset inesistente per un admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      },
      "delete": {
        "tags": [
          "trash"
        ],
        "summary": "Delete an asset",
        "operationId": "assets_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id dell'asset",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAssetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Azione applicata, riga di audit registrata"
          },
          "400": {
            "description": "disk_action non riconosciuto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Asset non visibile, o purged senza i permessi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Errore del database o del filesystem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
...
    "/api/v1/assets/{id}/restore": {
      "post": {
        "tags": [
          "trash"
        ],
        "summary": "Restore an asset from trash",
        "operationId": "assets_restore",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id dell'asset",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "File ripristinato al percorso originale"
          },
          "401": {
            "description": "Non autenticato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Asset non visibile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Niente da ripristinare, o percorso occupato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Errore del database o del filesystem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
...
    "/api/v1/trash": {
      "get": {
        "tags": [
          "trash"
        ],
        "summary": "List trash assets",
        "operationId": "trash_list",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "description": "Keyset deleted_at|id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "1..=100, default 50",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pagina del cestino navigabile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrashListPage"
                }
              }
            }
          },
          "400": {
            "description": "Cursore illeggibile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Errore del database",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/trash/empty": {
      "post": {
        "tags": [
          "trash"
        ],
        "summary": "Empty the trash",
        "operationId": "trash_empty",
        "responses": {
          "200": {
            "description": "Cestino svuotato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyTrashResponse"
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Solo owner/admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Errore del database o del filesystem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssetView": {
        "type": "object",
        "required": [
          "id",
          "folder_id",
          "filename",
          "size_bytes",
          "kind",
          "status"
        ],
        "properties": {
          "content_hash": {
            "type": [
              "string",
              "null"
            ]
          },
          "filename": {
            "type": "string"
          },
          "folder_id": {
            "type": "string"
          },
          "height": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "location": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GeoPointView"
              }
            ]
          },
          "place_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "taken_at_utc": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "thumbhash": {
            "type": [
              "string",
              "null"
            ]
          },
          "width": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
...
      "DeleteAssetRequest": {
        "type": "object",
        "required": [
          "disk_action"
        ],
        "properties": {
          "disk_action": {
            "type": "string",
            "description": "`kept`, `moved_to_trash`, o `purged` — nessun default: il client deve\nsempre scegliere (spec §6).",
            "example": "moved_to_trash"
          }
        }
      },
...
      "EmptyTrashResponse": {
        "type": "object",
        "required": [
          "emptied"
        ],
        "properties": {
          "emptied": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "GeoPointView": {
        "type": "object",
        "required": [
          "lat",
          "lon"
        ],
        "properties": {
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "Errore in formato RFC 9457. Il campo `type` è un codice stabile su cui i\nclient possono ramificare; `title` è in inglese e serve al debug, non\nall'utente finale — la traduzione avviene nel frontend.\n\nÈ anche uno schema del documento `OpenAPI`: è la forma su cui il client\nmobile ramifica, quindi va descritta, non solo servita.",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "example": "keeppix/unauthenticated"
          }
        }
      },
...
      "TrashItemView": {
        "type": "object",
        "required": [
          "id",
          "asset_id",
          "deleted_at",
          "original_path",
          "disk_action",
          "days_remaining"
        ],
        "properties": {
          "asset_id": {
            "type": "string"
          },
          "days_remaining": {
            "type": "integer",
            "format": "int64"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time"
          },
          "disk_action": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "original_path": {
            "type": "string"
          },
          "trash_path": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "TrashListPage": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrashItemView"
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-kpx_session",
        "description": "Cookie di sessione emesso da POST /api/v1/setup e POST /api/v1/auth/login."
      }
    }
  },
...
}